requirements.txt
beautifulsoup4
lxml
pillow
mecab-python3
gensim
